requirements.txt
streamlit==1.35.0
opencv-python-headless==4.9.0.80
numpy==1.23.5                # Downgraded to be compatible with TensorFlow 2.12.0
pillow==10.2.0
mtcnn==0.1.1
deepface==0.0.79
scikit-learn==1.3.2
tensorflow==2.12.0
<<<<<<< HEAD
protobuf<5,>=3.20
=======
protobuf==4.25.3             # Optional but safer to pin it explicitly
>>>>>>> 6f19a06 (Commit)
